Tidy up and add module for matcher tests
