+++ spacy/tests/matcher/test_matcher_logic.py
@@ -4,6 +4,7 @@ from __future__ import unicode_literals
 import pytest
 import re
 from spacy.matcher import Matcher
+from spacy.tokens import Doc
 
 
 pattern1    = [{'ORTH':'A', 'OP':'1'}, {'ORTH':'A', 'OP':'*'}]
@@ -55,7 +56,7 @@ def test_greedy_matching(doc, text, pattern, re_pattern):
     (pattern3, re_pattern3),
     (pattern4, re_pattern4),
     (pattern5, re_pattern5)])
-def test_match_consuming(doc,text, pattern, re_pattern):
+def test_match_consuming(doc, text, pattern, re_pattern):
     """Test that matcher.__call__ consumes tokens on a match similar to
     re.findall."""
     matcher = Matcher(doc.vocab)
@@ -63,3 +64,53 @@ def test_match_consuming(doc,text, pattern, re_pattern):
     matches = matcher(doc)
     re_matches = [m.span() for m in re.finditer(re_pattern, text)]
     assert len(matches) == len(re_matches)
+
+
+def test_operator_combos(en_vocab):
+    cases = [
+        ('aaab', 'a a a b', True),
+        ('aaab', 'a+ b', True),
+        ('aaab', 'a+ a+ b', True),
+        ('aaab', 'a+ a+ a b', True),
+        ('aaab', 'a+ a+ a+ b', True),
+        ('aaab', 'a+ a a b', True),
+        ('aaab', 'a+ a a', True),
+        ('aaab', 'a+', True),
+        ('aaa', 'a+ b', False),
+        ('aaa', 'a+ a+ b', False),
+        ('aaa', 'a+ a+ a+ b', False),
+        ('aaa', 'a+ a b', False),
+        ('aaa', 'a+ a a b', False),
+        ('aaab', 'a+ a a', True),
+        ('aaab', 'a+', True),
+        ('aaab', 'a+ a b', True)
+    ]
+    for string, pattern_str, result in cases:
+        matcher = Matcher(en_vocab)
+        doc = Doc(matcher.vocab, words=list(string))
+        pattern = []
+        for part in pattern_str.split():
+            if part.endswith('+'):
+                pattern.append({'ORTH': part[0], 'OP': '+'})
+            else:
+                pattern.append({'ORTH': part})
+        matcher.add('PATTERN', None, pattern)
+        matches = matcher(doc)
+        if result:
+            assert matches, (string, pattern_str)
+        else:
+            assert not matches, (string, pattern_str)
+
+
+def test_matcher_end_zero_plus(en_vocab):
+    """Test matcher works when patterns end with * operator. (issue 1450)"""
+    matcher = Matcher(en_vocab)
+    pattern = [{'ORTH': "a"}, {'ORTH': "b", 'OP': "*"}]
+    matcher.add('TSTEND', None, pattern)
+    nlp = lambda string: Doc(matcher.vocab, words=string.split())
+    assert len(matcher(nlp('a'))) == 1
+    assert len(matcher(nlp('a b'))) == 2
+    assert len(matcher(nlp('a c'))) == 1
+    assert len(matcher(nlp('a b c'))) == 2
+    assert len(matcher(nlp('a b b c'))) == 3
+    assert len(matcher(nlp('a b b'))) == 3
